Move schema validation helpers from og.webactions to og.api.
