+++ opengever/webactions/storage.py
@@ -2,12 +2,12 @@ from BTrees.IOBTree import IOBTree
 from BTrees.OIBTree import OIBTree
 from BTrees.OOBTree import OOBTree
 from datetime import datetime
+from opengever.api.validation import validate_no_unknown_fields
+from opengever.api.validation import validate_schema
 from opengever.webactions.exceptions import ActionAlreadyExists
 from opengever.webactions.interfaces import IWebActionsStorage
 from opengever.webactions.schema import IPersistedWebActionSchema
 from opengever.webactions.schema import IWebActionSchema
-from opengever.webactions.validation import validate_no_unknown_fields
-from opengever.webactions.validation import validate_schema
 from persistent.mapping import PersistentMapping
 from plone import api
 from Products.CMFPlone.interfaces import IPloneSiteRoot
